+++ Gymondo/GymondoiOSTests/ExerciseListViewControllerTests.swift
@@ -103,44 +103,3 @@ final class ExerciseListViewControllerTests: XCTestCase {
         }
     }
 }
-
-private extension ExerciseListViewController {
-    func simulateUserInitiatedExerciseLoad() {
-        refreshControl?.simulatePullToRefresh()
-    }
-
-    var isShowingLoadingIndicator: Bool {
-        return refreshControl?.isRefreshing == true
-    }
-
-    func numberOfRenderedExercisesViews() -> Int {
-        return tableView.numberOfRows(inSection: exerciseItemsSection)
-    }
-
-    func exerciseItemView(at row: Int) -> UITableViewCell? {
-        let dataSource = tableView.dataSource
-        let index = IndexPath(row: row, section: exerciseItemsSection)
-        return dataSource?.tableView(tableView, cellForRowAt: index)
-    }
-
-    private var exerciseItemsSection: Int {
-        return 0
-    }
-}
-
-private extension ExerciseItemCell {
-
-    var nameText: String? {
-        return nameLabel.text
-    }
-}
-
-private extension UIRefreshControl {
-    func simulatePullToRefresh() {
-        allTargets.forEach { target in
-            actions(forTarget: target, forControlEvent: .valueChanged)?.forEach {
-                (target as NSObject).perform(Selector($0))
-            }
-        }
-    }
-}
