+++ Gymondo/GymondoiOSTests/Helpers/ExerciseItemCell+TestHelpers.swift
@@ -1,0 +1,16 @@
+//
+//  ExerciseItemCell+TestHelpers.swift
+//  GymondoiOSTests
+//
+//  Created by Ali Elsokary on 25/08/2023.
+//
+
+import UIKit
+@testable import GymondoiOS
+
+extension ExerciseItemCell {
+
+    var nameText: String? {
+        return nameLabel.text
+    }
+}
